pyproject.toml
[project]
name = "twitter-classification"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    { name = "viertel97", email = "janik.viertel@outlook.com" },
]
dependencies = [
    "presidio-analyzer[transformers]",
    "presidio-anonymizer==2.2.357",
    "pandas==2.2.3",
    "spacy==3.8.4",
    "openai==1.65.1",
    "tqdm>=4.67.1",
    "fastapi==0.115.11",
    "scikit-learn>=1.6.1",
    "python-dotenv>=1.0.1",
    "quarter-lib>=0.0.25",
    "uvicorn>=0.34.0",
    "python-multipart>=0.0.20",
    "pydantic==2.9.2",
    "pip>=25.0.1",
]
requires-python = "==3.11.*"
readme = "README.md"
license = { text = "MIT" }


[tool.pdm]
distribution = false
